+++ src/components/ConnectedTasksJobs/ConnectedTasksJobsButton.js
@@ -1,0 +1,12 @@
+import { ConnectedTasksJobsPlugin, connectedTasksJobsPropTypes } from './ConnectedTasksJobsPlugin';
+
+const ConnectedTasksJobsButton = props => (
+  <ConnectedTasksJobsPlugin
+    {...props}
+    componentType="ConnectedTasksJobsButton"
+  />
+);
+
+ConnectedTasksJobsButton.propTypes = connectedTasksJobsPropTypes;
+
+export default ConnectedTasksJobsButton;