+++ src/components/ConnectedTasksJobs/index.js
@@ -1,0 +1,2 @@
+export { default as ConnectedTasksJobsButton } from './ConnectedTasksJobsButton';
+export { default as ConnectedTasksJobsPane } from './ConnectedTasksJobsPane';